requirements.txt
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
pip==21.3.1
protobuf==3.19.1
pytz==2021.3
setuptools==58.1.0
Werkzeug==2.0.2
